Handle Errors from new Function of Discord Markdown with `#`

+++ events/servers/MessageCreate.js
@@ -9,12 +9,21 @@ module.exports = {
             if (message.author.bot) return;
             //TODO: check later
             if (message.channel.type === 1) return message.reply('Não irei responder a mensagens no privado. Por favor execute comandos em um servidor.')
-	    if (message.content.startsWith("# ")){
-		try {
-	 		return await message.delete();
-		}
-		catch {}
-	    }
+            
+            
+            if (message.content.startsWith("# ")) {
+                try {
+                    const member = await message.guild.members.fetch(message.author.id);
+                    if (!member.permissions.has(Discord.PermissionFlagsBits.Administrator)) return await message.delete();
+                } catch (error) {
+                    if (error.code === 10008) {
+                        return
+                    } else {
+                        console.error(colors.red('Error:', error));
+                    }
+                }
+            }
+            
 
             const CounterChannelId = process.env.CounterChannelId;
             const CounterAlternativeChannelId = process.env.CounterAlternativeChannelId;
